app/offre-ingestion/requirements.txt
requests==2.32.5
google-api-core==2.28.1
google-cloud-storage==3.7.0
google-cloud-bigquery==3.39.0
numpy==2.4.0
sentence-transformers==5.2.0
